+++ apps/custom_actions/schema_utils.py
@@ -1,6 +1,9 @@
 from copy import deepcopy
 
 from django.core.exceptions import ValidationError
+from langchain_community.tools import APIOperation
+from langchain_community.utilities.openapi import OpenAPISpec
+from pydantic import BaseModel
 
 
 def get_standalone_schema_for_action_operation(action_operation):
@@ -81,3 +84,29 @@ def resolve_references(openapi_spec: dict) -> dict:
         return data
 
     return resolve_ref(deepcopy(openapi_spec), "")
+
+
+class APIOperationDetails(BaseModel):
+    operation_id: str
+    description: str
+    path: str
+    method: str
+
+    def __str__(self):
+        return f"{self.method.upper()}: {self.description}"
+
+
+def get_operations_from_spec_dict(spec_dict) -> list[APIOperationDetails]:
+    spec = OpenAPISpec.from_spec_dict(spec_dict)
+    return get_operations_from_spec(spec)
+
+
+def get_operations_from_spec(spec) -> list[APIOperationDetails]:
+    operations = []
+    for path in spec.paths:
+        for method in spec.get_methods_for_path(path):
+            op = APIOperation.from_openapi_spec(spec, path, method)
+            operations.append(
+                APIOperationDetails(operation_id=op.operation_id, description=op.description, path=path, method=method)
+            )
+    return operations
